Revert "move stuff into routing_tables"
This reverts commit 33771edb0d04bc63e5d172fee840da21676d5108.
