+++ c_common/models/minimise/src/minimise.h
@@ -18,6 +18,123 @@
  * (entry_t is defined in `routing_table.h` but is described below).
  */
 
+typedef struct {
+
+    // Application ID to use to load the routing table. This can be left as `0'
+    // to load routing entries with the same application ID that was used to
+    // load this application.
+    uint32_t app_id;
+
+    //flag for compressing when only needed
+    uint32_t compress_only_when_needed;
+
+    // flag that uses the available entries of the router table instead of
+    //compressing as much as possible.
+    uint32_t compress_as_much_as_possible;
+
+    // Initial size of the routing table.
+    uint32_t table_size;
+
+    // Routing table entries
+    entry_t entries[];
+} header_t;
+
+/* entry_t is defined as:
+ *
+ *     typedef struct
+ *     {
+ *       uint32_t key;
+ *       uint32_t mask;
+ *       uint32_t route;   // Routing direction
+ *       uint32_t source;  // Source of packets arriving at this entry
+ *     } entry_t;
+ *
+ * The `source` field is used to determine if the entry could be replaced by
+ * default routing, it can be left blank if removing default entries is not to
+ * be used. Otherwise indicate which links will be used by packets expected to
+ * match the specified entry.
+ *
+ * NOTE: The routing table provided to this application MUST include all of the
+ * entries which are expected to arrive at this router (i.e., entries which
+ * could be replaced by default routing MUST be included in the table provided
+ * to this application).
+ *
+ * NOTE: The block of memory containing the header and initial routing table
+ * will be freed on exit by this application.
+ */
+
+//! \brief prints the header object for debug purposes
+//! \param[in] header: the header to print
+void print_header(header_t *header) {
+    log_info("app_id = %d", header->app_id);
+    log_info(
+        "compress_only_when_needed = %d",
+        header->compress_only_when_needed);
+    log_info(
+        "compress_as_much_as_possible = %d",
+        header->compress_as_much_as_possible);
+    log_info("table_size = %d", header->table_size);
+}
+
+//! \brief Read a new copy of the routing table from SDRAM.
+//! \param[in] table : the table containing router table entries
+//! \param[in] header: the header object
+void read_table(table_t *table, header_t *header) {
+    // Copy the size of the table
+    table->size = header->table_size;
+
+    // Allocate space for the routing table entries
+    table->entries = MALLOC(table->size * sizeof(entry_t));
+
+    // Copy in the routing table entries
+    spin1_memcpy((void *) table->entries, (void *) header->entries,
+            sizeof(entry_t) * table->size);
+}
+
+//! \brief Load a routing table to the router.
+//! \param[in] table: the table containing router table entries
+//! \param[in] app_id: the app id for the routing table entries to be loaded
+//! under
+//! \return bool saying if the table was loaded into the router or not
+bool load_routing_table(table_t *table, uint32_t app_id) {
+
+    // Try to allocate sufficient room for the routing table.
+    uint32_t entry_id = rtr_alloc_id(table->size, app_id);
+    if (entry_id == 0) {
+        log_info("Unable to allocate routing table of size %u\n", table->size);
+        return FALSE;
+    }
+
+    // Load entries into the table (provided the allocation succeeded).
+    // Note that although the allocation included the specified
+    // application ID we also need to include it as the most significant
+    // byte in the route (see `sark_hw.c`).
+    for (uint32_t i = 0; i < table->size; i++) {
+        entry_t entry = table->entries[i];
+        uint32_t route = entry.route | (app_id << 24);
+        rtr_mc_set(entry_id + i, entry.keymask.key, entry.keymask.mask,
+                route);
+    }
+
+    // Indicate we were able to allocate routing table entries.
+    return TRUE;
+}
+;
+//! \brief frees memory allocated and calls spin1 exit and sets the user0
+//! error code correctly.
+//! \param[in] header the header object
+//! \param[in] table the data object holding the routing table entries
+void cleanup_and_exit(header_t *header, table_t table) {
+    // Free the memory used by the routing table.
+    log_debug("free sdram blocks which held router tables");
+    FREE(table.entries);
+    // Free the block of SDRAM used to load the routing table.
+    sark_xfree(sv->sdram_heap, (void *) header, ALLOC_LOCK);
+
+    log_info("completed router compressor");
+    sark.vcpu->user0 = 0;
+    spin1_exit(0);
+}
 
 void minimise(table_t *table, uint32_t target_length);
